+++ examples/heightfield.rs
@@ -17,7 +17,7 @@ use futures_lite::future;
 use oxidized_navigation::{
     query::{find_path, find_polygon_path, perform_string_pulling_on_path},
     tiles::NavMeshTiles,
-    NavMesh, NavMeshAffector, NavMeshSettings, OxidizedNavigationPlugin,
+    NavMesh, NavMeshAffector, NavMeshSettings, OxidizedNavigationPlugin, debug_draw::{OxidizedNavigationDebugDrawPlugin, DrawNavMesh, DrawPath},
 };
 
 fn main() {
@@ -42,6 +42,7 @@ fn main() {
                     max_tile_generation_tasks: Some(9),
                 },
             },
+            OxidizedNavigationDebugDrawPlugin,
             // The rapier plugin needs to be added for the scales of colliders to be correct if the scale of the entity is not uniformly 1.
             // An example of this is the "Thin Wall" in [setup_world_system]. If you remove this plugin, it will not appear correctly.
             RapierPhysicsPlugin::<NoUserData>::default(),
@@ -51,7 +52,6 @@ fn main() {
             ..Default::default()
         })
         .insert_resource(AsyncPathfindingTasks::default())
-        .insert_resource(DrawNavMesh(false))
         .add_systems(Startup, (setup_world_system, info_system))
         .add_systems(
             Update,
@@ -63,7 +63,6 @@ fn main() {
                 spawn_or_despawn_affector_system,
             ),
         )
-        .add_systems(PostUpdate, draw_path)
         .run();
 }
 
@@ -105,7 +104,7 @@ fn run_blocking_pathfinding(
                     Ok(string_path) => {
                         info!("String path (BLOCKING): {:?}", string_path);
                         commands.spawn(DrawPath {
-                            timer: Timer::from_seconds(4.0, TimerMode::Once),
+                            timer: Some(Timer::from_seconds(4.0, TimerMode::Once)),
                             pulled_path: string_path,
                             color: Color::RED,
                         });
@@ -171,7 +170,7 @@ fn poll_pathfinding_tasks_system(
         if let Some(string_path) = future::block_on(future::poll_once(task)).unwrap_or(None) {
             info!("Async path task finished with result: {:?}", string_path);
             commands.spawn(DrawPath {
-                timer: Timer::from_seconds(4.0, TimerMode::Once),
+                timer: Some(Timer::from_seconds(4.0, TimerMode::Once)),
                 pulled_path: string_path.clone(),
                 color: Color::BLUE,
             });
@@ -215,75 +214,16 @@ async fn async_path_find(
     None
 }
 
-#[derive(Component)]
-struct DrawPath {
-    timer: Timer,
-    pulled_path: Vec<Vec3>,
-    color: Color,
-}
-
-// Helper function to draw a path for the timer's duration.
-fn draw_path(
-    mut commands: Commands,
-    mut path_query: Query<(Entity, &mut DrawPath)>,
-    time: Res<Time>,
-    mut gizmos: Gizmos,
-) {
-    for (entity, mut draw_path) in path_query.iter_mut() {
-        if draw_path.timer.tick(time.delta()).just_finished() {
-            commands.entity(entity).despawn();
-        } else {
-            gizmos.linestrip(draw_path.pulled_path.clone(), draw_path.color);
-        }
-    }
-}
-
-#[derive(Resource)]
-struct DrawNavMesh(bool);
-
 //
 //  Draw Nav-mesh.
-//  Press M to toggle drawing the navmesh..
-//
-//  This uses the crate ``bevy_prototype_debug_lines``.
+//  Press M to toggle drawing the navmesh.
 //
 fn draw_nav_mesh_system(
     keys: Res<Input<KeyCode>>,
-    nav_mesh: Res<NavMesh>,
     mut draw_nav_mesh: ResMut<DrawNavMesh>,
-    mut gizmos: Gizmos,
 ) {
     if keys.just_pressed(KeyCode::M) {
         draw_nav_mesh.0 = !draw_nav_mesh.0;
-    }
-
-    if !draw_nav_mesh.0 {
-        return;
-    }
-    if let Ok(nav_mesh) = nav_mesh.get().read() {
-        for (tile_coord, tile) in nav_mesh.get_tiles().iter() {
-            let tile_color = Color::Rgba {
-                red: 0.0,
-                green: (tile_coord.x % 10) as f32 / 10.0,
-                blue: (tile_coord.y % 10) as f32 / 10.0,
-                alpha: 1.0,
-            };
-            // Draw polygons.
-            for poly in tile.polygons.iter() {
-                let indices = &poly.indices;
-                for i in 0..indices.len() {
-                    let a = tile.vertices[indices[i] as usize];
-                    let b = tile.vertices[indices[(i + 1) % indices.len()] as usize];
-
-                    gizmos.line(a, b, tile_color);
-                }
-            }
-
-            // Draw vertex points.
-            for vertex in tile.vertices.iter() {
-                gizmos.line(*vertex, *vertex + Vec3::Y, tile_color);
-            }
-        }
     }
 }